requirements.txt
fastapi[all]
python-dotenv
aiohttp
langchain[openai]
beanie